+++ java/src/main/resources/db/migration/V1__create_tables.sql
@@ -1,14 +1,14 @@
 CREATE TABLE time_deposits (
-                            id BIGSERIAL PRIMARY KEY,
+                            id SERIAL PRIMARY KEY,
                             plan_type VARCHAR(50) NOT NULL,
                             days INTEGER NOT NULL,
-                            balance DOUBLE PRECISION NOT NULL
+                            balance DECIMAL(15, 2) NOT NULL
 );
 
 CREATE TABLE withdrawals (
-                            id BIGSERIAL PRIMARY KEY,
+                            id SERIAL PRIMARY KEY,
                             time_deposit_id INTEGER NOT NULL,
-                            amount DOUBLE PRECISION NOT NULL,
+                            amount DECIMAL(15, 2) NOT NULL,
                             transaction_date DATE NOT NULL,
 
                             CONSTRAINT fk_deposit
